+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Blackcube Core - CHANGELOG 
 ==========================
 
+ * Upd: Update dependencies
+ * Upd: Update `Element` to handle bloc extraction
+ * Upd: Handle canonical URL in `SeoBehavior`
+
 Release 3.0.3 October, 4th 2023 <pgaultier@redcat.io>
 ----------------------------------------------------